refactor: Extract network menu methods to mixin (1,532 → 1,404 lines)
- Move _network_menu and _run_terminal_network to network_tools_mixin.py
- Reduces launcher_tui/main.py from 1,532 to 1,404 lines (8.4% reduction)
- All files now under 1,500 line threshold per Issue #6 guidelines
- Update CLAUDE.md and SESSION_NOTES.md with current file sizes

Files changed:
- src/launcher_tui/network_tools_mixin.py: Added network menu methods
- src/launcher_tui/main.py: Removed extracted methods
